+++ package.json
@@ -23,6 +23,7 @@
   "devDependencies": {
     "@medic/eslint-config": "^1.1.0",
     "chai": "^4.2.0",
+    "chai-exclude": "^2.0.2",
     "enketo-core": "4.41.6",
     "eslint": "^6.8.0",
     "jsdoc": "^3.6.3",
